split: sort split files

+++ pkg/unit.go
@@ -1,4 +1,4 @@
 package pkg
 
-const Page = int64(4096)
+const Page = 4096
 const GB = 1_073_741_824
